Refactor - move PieChart code out to own module
* I also cleaned up the charting code inc drawing on clicking
'recent measurements'
