+++ src/app/pie-chart.spec.ts
@@ -1,0 +1,7 @@
+import { PieChart } from './pie-chart';
+
+describe('PieChart', () => {
+  it('should create an instance', () => {
+    expect(new PieChart()).toBeTruthy();
+  });
+});
